create VideoTileCover
